refactor: split loader validation and normalization
